requirements.txt
fastapi
uvicorn
requests
python-jose
passlib
python-multipart
asyncio
motor
tinytag
httpx